split all fixtures
